requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Serial communication
pyserial>=3.5

# Visualization (required for validation module)
matplotlib>=3.5.0

# Optional: For advanced data analysis
# pandas>=1.3.0
